Extract platform-neutral worklet client
