+++ src/rust/shoop_worklet_client/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "shoop_worklet_client"
+version.workspace = true
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+serde_json = { workspace = true }
+shoop_app_api = { path = "../shoop_app_api" }
+shoop_audio_protocol = { path = "../shoop_audio_protocol" }
+shoop_backend = { path = "../shoop_backend" }
+
+[dev-dependencies]
+tracy-nextest-capture = { workspace = true }
+
+[lints]
+workspace = true
