+++ src/rust/shoopdaloop/Cargo.toml
@@ -32,6 +32,7 @@ log = { workspace = true }
 serde_json = { workspace = true }
 shoop_audio_protocol = { path = "../shoop_audio_protocol" }
 shoop_session = { path = "../shoop_session" }
+shoop_worklet_client = { path = "../shoop_worklet_client" }
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
 web-time = "1.1"
